Check for storage existance before trying to check.

+++ src/modloader/modloader64.ts
@@ -74,18 +74,20 @@ class ModLoader64 {
         process.emit('SIGINT');
       });
 
-      fs.readdirSync('./storage').forEach((key: string) => {
-        let file: string = path.join('./storage', key);
-        if (fs.existsSync(file)) {
-          let seconds =
-            (new Date().getTime() -
-              new Date(fs.statSync(file).mtime).getTime()) /
-            1000;
-          if (seconds > 2592000) {
-            fs.unlinkSync(file);
+      if (fs.existsSync('./storage')) {
+        fs.readdirSync('./storage').forEach((key: string) => {
+          let file: string = path.join('./storage', key);
+          if (fs.existsSync(file)) {
+            let seconds =
+              (new Date().getTime() -
+                new Date(fs.statSync(file).mtime).getTime()) /
+              1000;
+            if (seconds > 2592000) {
+              fs.unlinkSync(file);
+            }
           }
-        }
-      });
+        });
+      }
     }
 
     process.on('SIGINT', function() {
